mod/#8 - ProductDetail 관련 컴포넌트 위치 변경 및 model nullable 반영
